Don't let Travis CI delete source JAR files during deployment
I decided to delete the source JAR files in build artifacts because
users might be confused on which file they should download. However, I
also noticed that GitHub only adds archives of source code on the
branch where I created the tag linked to a release. This means source
code of other branches cannot be downloaded from the Releases page.
I thought this was not an issue because users who would like to see
source code of other branches could still find it on GitHub. But what
if Hypixel stops supporting an old Minecraft version, and we delete the
branch holding code for that version? Users cannot find source code for
that version, which is not what we want to see.
I finally decide to keep the source JAR files generated during
deployment and add some explaining words in README.md. Hopefully, it
will guide the users to download the correct file.

(cherry picked from commit c5707cb)