Completely removes scoreboard reading in GameDetector.java
Scoreboard tells information about a Minecraft multi-player mini game,
such as time until the next event in game, teams, etc. Scoreboard also
tells us something else: in Hypixel, different scoreboards are shown
when you are in Bed Wars lobby, when you wait for a game to start, and
when you are playing Bed Wars.
So, when I thought about how to write this mod for the first time, I
decided to detect if the client is in a Bed Wars game by reading
scoreboard. I also planned to do other things by reading scoreboard,
such as getting time till next diamond/emerald by reading how much is
left until the next generator upgrade, and getting number of teams.
However, I disappointingly found that reading scoreboard is unreliable:
it stores more than what you see. You see a scoreboard with size 15 in
a Bed Wars game; however, because of hidden entries that are not shown
on the scoreboard, there are actually more than 40 lines.
I happened to find that when client rejoins a Bed Wars game, there is
always a chat message telling you to type `/lobby` if you want to leave
a game. I thought, since this message is always received upon rejoining
a game, we can stop checking scoreboard to see if client is entering a
game or a lobby and read that message instead.