Add ability to show team upgrades the player has got
I did not implement tracking of "Sharpened Sword", "Reinforced Armor",
and "Maniac Miner" because they can be inferred from appearance of the
player's sword, Protection enchantment level of the player's armor, and
status effect on the player.
I would also like to track traps, but because Hypixel is changing them
in Bed Wars v1.4, and I am not so familiar with the new trap system yet,
traps tracking is pending and will be added in future updates.