Move all decisions of displayed strings and icons to GuiHud
GuiHud is the class responsible for HUD GUI rendering, which should
include the format of diamond/emerald generation time, text displayed
when there is no available generator to read, and which icon represents
each resource generation level and trap. The purpose of GameManager is
solely providing information related to the current game session, e.g.
whether there is a generator available, time until the next generation,
which traps are in the current trap queue.

(cherry picked from commit 2de611d)