Add update checking using Forge's built-in update checker
Minecraft Forge itself has an update checker which is used for
notifying its users when an update is available by showing a flashing
emerald-like icon on the "Mods" button in the main menu. Luckily,
Minecraft Forge mods can also use this update checker. For more
technical details, refer to
https://mcforge.readthedocs.io/en/latest/gettingstarted/autoupdate/.

To make this mod's version number compatible with the update checker, I
had to remove the supported Minecraft client versions from the version
number. A special thing about HBW Helper's JARs is that it may support
more than one client version, and I had indicated what versions are
supported by adding something like [1.11, 1.12.2] to the version number.
This does not work with Minecraft Forge update checker, so I removed it.
However, the supported versions of a JAR would still be indicated in
the JAR's file name, so it should not be a big deal.

(cherry picked from commit 0290dde)