Modify trap mechanism for Bed Wars Castle mode
Because a trap can be used for multiple times in Bed Wars Castle mode,
this commit was made to support the new variant of trap system in
Hypixel Bed Wars.

As of this commit was made, there were still some limitations:
- Now, we have to rely on the unreliable scoreboard again to get the
  current game mode of Bed Wars. If the scoreboard were not fully loaded
  when we would try to read it, detection of game mode would fail.
- In Bed Wars Castle mode, because a trap can be consumed for multiple
  times, the first trap in the trap queue might not be the one being set
  off. This would lead to incorrect reflection of the local trap queue.
These limitations might be resolved later.

(cherry picked from commit 77a28a0)