Improve game type detection mechanism
The former implementation attempts to detect the game type by reading
text on the scoreboard x seconds after a game starts, where x is a fixed
number. We wish x to be small so that a new GameManager instance for
the game type can be created as soon as possible. However, due to
network connection quality and computer performance, a fixed value of x
might not always work well: if Hypixel lags, or the user's computer is
very slow, scoreboard might not be fully loaded after x seconds, nor
will game type detection be successful.

This new implementation keeps monitoring the scoreboard after the game
starts until it has successfully detected the game type. It is like
naturally changing the value of x to the exact time length we desire.
The issue described above would be resolved by this implementation.