Prevent GameManager instance from being unnecessarily cleared
When client is already in a Bed Wars game and joins an in-progress game,
the current GameManager instance should be cleared and discarded, and a
new one should be created. After the client receives the "Found an
in-progress Bed Wars game" message, chances are the client is not sent
to that in-progress game because the current game and the in-progress
game are the same, or the in-progress game has been closed to new
players, which results the client still being in the current game.
However, if the GameManager instance is cleared when that message is
received, and the client does not leave the current game because of the
reason above, the game continues without game information display, which
is not what we want. This commit resolves this issue by clearing
GameManager instance only when client really switches from one game to
another.