Allow the user to select the current Dream mode
This reduces the need to update this mod when Hypixel change which
Dream mode game runs on the server. When Hypixel introduce a brand new
Dream mode, it is still necessary to update the mod; however, as they
have promised, Hypixel also rotate old Dream mode games, so when an
existing Dream mode game is reintroduced, this feature would be very
helpful.