Improve performance on showing prompts in chat
Rather than creating a TextComponentString/ChatComponentText object
every time when a message is being sent to the user, now the objects
are cached during class initialization, decreasing times of object
creation during runtime.

The drawback of this commit is that if the user changes the language of
the Minecraft client, the prompts will still be shown in the previous
selected language. However, most users would not change the language
more frequently than the prompts of this mod being shown.

(cherry picked from commit c80bbe7)

(cherry picked from commit 2ced1e5)