Modify GameManager for safe object construction
According to Section 3.2.1 in Java Concurrency in Practice by Goetz et
al., even if we publish the newly-created GameManager object in the last
statement of the constructor, other programs might still see an
improperly-constructed object. To resolve this issue, the new static
method is introduced for moving the publication statement out from the
constructor.