Fix improper LinkedHashMap initial capacity during construction
I did not know until recently that the initial capacity for
hash table-based collections is not equivalent to expected size of it.
Fortunately, Google's Guava library includes code that converts
expected size to initial capacity. Because Guava has been declared as a
dependency in Forge MDK, we can use this library in mod development
without extra effort.