Fix improper LinkedHashMap initial capacity during construction
I did not know until recently that the initial capacity for
hash table-based collections is not equivalent to expected size of it.
Although Google's Guava library includes code that converts
expected size to initial capacity, the version of Guava declared by
Forge MDK for these Minecraft client versions does not have method for
creating a pre-sized LinkedHashMap, so I would rather use the default
capacity, which is sufficient for a few elements and would not rehash.