Update CONTRIBUTING.md
I do not like using the word "code" as a verb. It sounds unprofessional.
"Programming" and "programmer" are good words with the same meaning as
"coding" and "coder" respectively and have longer history than the
latter words. But why did I use it so many times before? Nevertheless, I
should change it, even if this document might have not been carefully
read by anyone.

[skip ci]