Port this mod to 1.14.4
The only thing in this mod affected by the massive changes in Minecraft
and Minecraft Forge code base since 1.12.2 is the lack of a user
interface for changing this mod's configuration. Forge has removed the
bundled configuration GUI utilities, probably because the framework does
not work with the new Minecraft versions.

I believe the developers of Forge have a plan to create a new set of
configuration utilities, but the new framework is not here yet, so I am
planning to build a minimal user interface for changing this mod's
configuration and use it until Forge makes any progress. The new
interface is likely to be a CLI because I do not have enough time to
learn and build a GUI for this purpose.