Add command for changing this mod's configuration
This is just a temporary solution to the lack of a configuration GUI
framework in Minecraft Forge 28.0. If a new framework is added back to
Forge later, then this command line interface should be at least
deprecated because a configuration GUI is always preferred to a CLI. For
the same reason, this command should not be back-ported to branches for
older Minecraft client versions: it is only for 1.14.4 and any future
version which does not have configuration GUI support from Forge.