Deploy to GitHub Pages before GitHub Releases
In the most recent deployment of this mod, the JSON for Forge update
checker was wrongly updated again. There was an attempt to fix a similar
occurrence of this issue before, but the last deployment has proven that
it was not enough, so I hope this additional change could completely
solve the problem.