Defer writes to configuration file until ConfigScreen closes
If a configuration change is saved immediately after it is made, and
the user clicks on the option for it too quickly, the value shown for
the option on the GUI might not update correctly because saving the
change to file might take a little time. To optimize for user
experience, we could transiently keep the updated configuration in
memory, then commit the changes when the settings screen is closed.