Use nil UUID when sending messages to the player
The method for sending a message to the player in Minecraft has started
to require an extra UUID argument since 1.16.  On the client side, the
argument is actually unused, so any value with compatible type works.
However, when Minecraft itself calls the method, it uses Util.NIL_UUID
when the UUID is irrelevant, so it is probably a good idea for mods to
follow the same practice too.